Move commands inot a separate module
