+++ prio/__main__.py
@@ -1,0 +1,5 @@
+from .cli import main
+
+if __name__ == "__main__":
+    import sys
+    sys.exit(main())
